+++ src/main/kotlin/com/jspl/tickettaka/notification/AuthenticationService.kt
@@ -2,9 +2,9 @@ package com.jspl.tickettaka.notification
 
 import org.springframework.stereotype.Service
 
-@Service
-class AuthenticationService(private val memberRepository: MemberRepository) {
-    fun getMemberOrThrow(key: String): Member {
-        return memberRepository.findByKey(key) ?: throw MemberNotFoundException("Member not found for key: $key")
-    }
-}
+//@Service
+//class AuthenticationService(private val memberRepository: MemberRepository) {
+//    fun getMemberOrThrow(key: String): Member {
+//        return memberRepository.findByKey(key) ?: throw MemberNotFoundException("Member not found for key: $key")
+//    }
+//}
