+++ src/main/kotlin/com/jspl/tickettaka/notification/NotificationEventHandler.kt
@@ -4,11 +4,11 @@ import org.springframework.context.event.EventListener
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
 
-@Component
-class NotificationEventHandler(private val notificationService: NotificationService) {
-    @Async
-    @EventListener // 이벤트 구독
-    fun handleEvent(event: NotificationEvent) {
-        notificationService.sendNotification(event)
-    }
-}
+//@Component
+//class NotificationEventHandler(private val notificationService: NotificationService) {
+//    @Async
+//    @EventListener // 이벤트 구독
+//    fun handleEvent(event: NotificationEvent) {
+//        notificationService.sendNotification(event)
+//    }
+//}
