--- src/main/kotlin/com/jspl/tickettaka/notification/NotificationRepository.kt
@@ -1,6 +1,0 @@
-package com.jspl.tickettaka.notification
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface NotificationRepository : JpaRepository<NotificationEvent, Long> {
-}
