+++ src/main/kotlin/com/jspl/tickettaka/repository/NotificationRepository.kt
@@ -1,0 +1,7 @@
+package com.jspl.tickettaka.repository
+
+import com.jspl.tickettaka.model.Notification
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface NotificationRepository : JpaRepository<Notification, Long> {
+}
