+++ src/main/kotlin/com/jspl/tickettaka/service/EmitterService.kt
@@ -1,5 +1,6 @@
-package com.jspl.tickettaka.notification
+package com.jspl.tickettaka.service
 
+import com.jspl.tickettaka.repository.EmitterRepository
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter
@@ -7,20 +8,20 @@ import java.util.concurrent.ConcurrentHashMap
 import java.util.*
 
 @Service
-class EmitterService(private val emitterRepository: EmitterRepository) {
+class EmitterService: EmitterRepository {
     private val emitters: MutableMap<String, SseEmitter> = ConcurrentHashMap()
     private val log = LoggerFactory.getLogger(EmitterService::class.java)
 
-    fun save(eventId: String, sseEmitter: SseEmitter): SseEmitter {
+    override fun save(eventId: String, sseEmitter: SseEmitter): SseEmitter {
         emitters[eventId] = sseEmitter
         return sseEmitter
     }
 
-    fun find(memberId: String): Optional<SseEmitter> {
+    override fun findById(memberId: String): Optional<SseEmitter> {
         return Optional.ofNullable(emitters[memberId])
     }
 
-    fun delete(eventId: String) {
+    override fun deleteById(eventId: String) {
         emitters.remove(eventId)
     }
 }
